Linear_SVM_Classification
A classification model using LinearSVC was built with the Iris dataset. Petal length and petal width features were used to predict the Iris Virginica class. The model was scaled using StandardScaler, with the C parameter set to 1. Additionally, the decision_function was applied to calculate the distances of predictions from the decision boundary.